+++ tests/testthat.R
@@ -3,3 +3,4 @@ library(tidyverse)
 library(dtrackr)
 
 test_check("dtrackr")
+